requirements.txt
Django==5.0.1
django_cors_headers==4.3.1
django_environ==0.11.2
djangorestframework==3.14.0
